refactor: separate notification and stock services
