+++ app/services/notification.py
@@ -1,13 +1,11 @@
 import asyncio
 import logging
-import random
 from asyncio import CancelledError
 from datetime import datetime
 from decimal import Decimal
 from typing import List, Optional
 from uuid import uuid4
 
-import httpx
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from fastapi import HTTPException
 from pydantic import ValidationError
@@ -21,12 +19,11 @@ from app.models.models import (StockPriceNotificationCreateRq,
                                StockPriceNotificationReadRq,
                                TelegramUser,
                                StockPriceNotificationDeleteRq,
-                               StockRs,
-                               StockRq, ExchangeSuffix, ExchangeRs, FindStockRq, Amount, ActionsOnExchange,
+                               StockRq, ActionsOnExchange,
                                NotificationMessage)
+from app.services.stock import StockService
 
 # todo: add some tests
-
 setup_logging()
 logger = logging.getLogger(__name__)
 
@@ -329,112 +326,4 @@ class NotificationStockPriceService:
         raise NotImplementedError
 
     async def delete_many(self, user: TelegramUser):
-        raise NotImplementedError
-
-
-class YahooApiService:
-    """
-    Base class for calling YahooApi
-    """
-    __user_agent_lst = ['Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) '
-                        'Chrome/88.0.4324.104 Safari/537.36',
-                        'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:84.0) Gecko/20100101 Firefox/84.0',
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_0) AppleWebKit/537.36 (KHTML, like Gecko) '
-                        'Chrome/75.0.3770.100 Safari/537.36']
-    __headers = {
-        'authority': 'query1.finance.yahoo.com',
-        'user-agent': random.choice(__user_agent_lst)
-    }
-
-    @classmethod
-    async def fetch_data(cls, url: str):
-        """
-        Get response from API by url
-        :param url: url
-        :return: dict with data from API
-        """
-
-        try:
-            async with httpx.AsyncClient() as client:
-                r = await client.get(url, headers=cls.__headers)
-                response = r.json()
-                logger.info(f'Response from YahooFinance: {r.status_code}')
-                r.raise_for_status()
-            return response
-        except httpx.HTTPError as exc:
-            logger.warning(f'HTTP Exception: {exc}')
-
-
-class StockService(YahooApiService):
-    """
-    Base class for stock
-    """
-    module = 'price'
-
-    async def find_stocks_by_ticker(self, stock: FindStockRq) -> Optional[List[StockRs]]:
-        """
-        Функция по тикеру возвращает представление
-        инструмента на каждой из бирж
-
-        :param stock: (ticker - тикер)
-        :return: список представлений инструмента на каждой из бирж
-        """
-        logger.info(f'Start checking exchange!')
-
-        response_list = []
-        try:
-            yahoo_symbol_list = [stock.ticker + i for i in ExchangeSuffix]
-            urls = [f'https://query1.finance.yahoo.com/v10/finance/quoteSummary/'
-                    f'{s}?modules={self.module}' for s in yahoo_symbol_list]
-            tasks = [asyncio.create_task(asyncio.wait_for(self.fetch_data(url), timeout=3)) for url in
-                     urls]
-            result_lst = [r.get("quoteSummary")["result"][0]["price"] for r in await asyncio.gather(*tasks) if r]
-
-            for item in result_lst:
-                exchange = ExchangeRs(
-                    code=item.get("exchange"),
-                    name=item.get("exchangeName"),
-                    yahoo_search_symbol=item.get("symbol")
-                )
-                amount = Amount(
-                    value=item["regularMarketPrice"].get("raw"),
-                    currency=item.get("currency"),
-                    currency_symbol=item.get("currencySymbol")
-                )
-                asset = StockRs(
-                    shortName=item.get("shortName"),
-                    price=amount,
-                    ticker=item.get("symbol").partition('.')[0],
-                    exchange=exchange
-                )
-                response_list.append(asset)
-            logger.debug(f'Returning objects: {response_list}')
-            return response_list
-        except CancelledError:
-            done, pending = await asyncio.wait(asyncio.tasks.all_tasks())
-            await asyncio.gather(pending)
-
-    async def get_stock_price(self, stock: StockRq) -> Amount:
-        """
-        Getting stock price from YahooApi
-        :param stock: model: StockRq
-        :return: model: Amount
-        """
-        try:
-            url = f'https://query1.finance.yahoo.com/v10/finance/quoteSummary/' \
-                  f'{stock.exchange.yahoo_search_symbol}?modules={self.module}'
-            response = await asyncio.create_task(asyncio.wait_for(self.fetch_data(url), timeout=3))
-            price = response.get("quoteSummary")["result"][0]["price"]
-            amount = Amount(
-                value=price["regularMarketPrice"].get("raw"),
-                currency=price.get("currency"),
-                currency_symbol=price.get("currencySymbol")
-            )
-            logger.debug(f'Return amount {amount}')
-            return amount
-        except CancelledError:
-            done, pending = await asyncio.wait(asyncio.tasks.all_tasks())
-            await asyncio.gather(pending)
-
-    async def stock_profile(self, stock: StockRq):
         raise NotImplementedError
